% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gender_ratio.R
\name{calc_gender_ratio}
\alias{calc_gender_ratio}
\title{Create the gender ratio plot}
\usage{
calc_gender_ratio(input_stat, input_event)
}
\arguments{
\item{input_stat}{Shiny user-interface input for the selected statistic}

\item{input_event}{Shiny user-interface input for the selected event date}
}
\description{
This function receives reactive inputs from the US PhD diversity R shiny app
and creates an interactive plot to visualise the selected statistic
}
