% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gender_ratio0.R
\name{calc_gender_ratio0}
\alias{calc_gender_ratio0}
\title{Create the gender ratio plot without using the R shiny application}
\usage{
calc_gender_ratio0(input_stat, input_event)
}
\arguments{
\item{input_stat}{Input for the selected statistic. The selected statistics can be "Gender_ratio","Prop_Male" and "Prop_Female"}

\item{input_event}{Input for the selected event period. The event periods can be "Pre 2000" or "Post 2000"}
}
\description{
This function receives inputs similar to those from the US PhD diversity R shiny app
and creates an interactive plot to visualise the selected statistic for the chosen time period
}
