% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_grads.R
\name{calc_grads}
\alias{calc_grads}
\title{Calculate the number of PhD graduates in the respective field}
\usage{
calc_grads(x)
}
\arguments{
\item{x}{The field of research for which, we are interested to calculate the number of PhD graduates}
}
\value{
The number of PhDs who have successfully graduated in a given field of study
}
\description{
The function calculates the total number of PhD graduates for the input field of study present in the database. The input field must be a valid PhD field currently practiced in the US.
}
\examples{
calc_grads("Agricultural economics")
}
