% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_app.R
\name{grad_app}
\alias{grad_app}
\title{US PhD diversity  R Shiny Application}
\usage{
grad_app()
}
\value{
A Shiny App embedded in the package
}
\description{
Run US PhD field and gender diversity R Shiny Application from inst/myapp directory
}
