\docType{package}
\name{gradr-package}
\alias{gradr}
\alias{gradr-package}
\title{gradr: Exploring the vast diversity of the US PhD graduates}
\description{
The gradr package is designed to study and explore the diversity among the PhD graduates in The United States.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ETC5523-2023/rpkg-arinbaruah.git }
  \item Report bugs at \url{Report bugs at https://github.com/ETC5523-2023/rpkg-arinbaruah/issues}
}

}
\author{
\strong{Maintainer}: Arindom Baruah \email{abar0090@student.monash.edu}

}
\keyword{internal}
