% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_phd_field.R
\docType{data}
\name{us_phd_field}
\alias{us_phd_field}
\title{US PhD field of study diversity}
\format{
A data frame with 3370 rows and 5 variables.
\describe{
\item{broad_field}{Broad field of study}
\item{major_field}{Major field of study within a broad field}
\item{field}{Specific field of study within a major and a broad field}
\item{year}{The year corresponding to the available data}
\item{n_phds}{The number of PhD graduates for the corresponding year and field of study}
}
}
\source{
\url{https://github.com/rfordatascience/tidytuesday/blob/master/data/2019/2019-02-19}
}
\usage{
us_phd_field
}
\description{
This dataset contains the information of the various fields of study currently
undertaken by the PhD graduates in the US. The fields of study have been classified
into "broad field", "major field" and "field" to introduce further granularity in the data.
The data encompasses all the number of PhD graduates in the US between the period of 2008-2017.
}
\examples{
head(us_phd_field)

}
\keyword{datasets}
