% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_phd_gender.R
\docType{data}
\name{us_phd_gender}
\alias{us_phd_gender}
\title{US PhD gender diversity}
\format{
A data frame with 53 rows and 4 variables.
\describe{
\item{Year}{The academic year of the gender diversity records}
\item{Male}{The number of male PhD graduates (in 1000s) for the corresponding academic year}
\item{Female}{The number of female PhD graduates (in 1000s) for the corresponding academic year}
\item{History}{An indicator to classify whether the academic year is post or pre 2000}
}
}
\source{
\url{https://www.statista.com/statistics/185167/number-of-doctoral-degrees-by-gender-since-1950/}
}
\usage{
us_phd_gender
}
\description{
This dataset contains the information pertaining to the gender diversity among the PhD graduates in the US.
The data encompasses the number of graduates between the time period of 1950-2020.
}
\examples{
head(us_phd_gender)
}
\keyword{datasets}
